Mulcloud API v4.5.5/requirements.txt
requests==2.21.0
Pillow==8.3.2
